chore: move footer to shared package
