+++ packages/apps/components/AppLayout.tsx
@@ -4,11 +4,11 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { PropsWithChildren, useMemo, useState } from 'react';
+import { Footer } from 'shared/components/Footer';
 import { Icon } from 'shared/components/widget/Icon';
 import { THEME } from 'shared/config/theme';
 import { readStorage } from 'shared/utils/helper';
 import { useITranslation } from '../hooks';
-import { Footer } from './Footer';
 import { Navigator } from './Navigator';
 import { Tools } from './Tools';
 import { ActiveAccount } from './widget/account/ActiveAccount';
