+++ packages/apps/hooks/index.ts
@@ -1,4 +1,4 @@
+export * from 'shared/hooks/translation';
 export * from './allowance';
 export * from './checkSpecVersion';
-export * from './translation';
 export * from './tx';
