+++ packages/helix/components/AppLayout.tsx
@@ -4,10 +4,10 @@ import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { PropsWithChildren, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { Footer } from 'shared/components/Footer';
 import { THEME } from 'shared/config/theme';
 import { readStorage } from 'shared/utils/helper';
 import { Path } from '../config';
-import { Footer } from './Footer';
 import { Navigator } from './Navigator';
 
 const { Header, Content } = Layout;
