+++ packages/shared/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './darwiniaBalances';
 export * from './localSearch';
 export * from './isMounted';
+export * from './translation';
