+++ packages/shared/package.json
@@ -28,14 +28,15 @@
     "@polkadot/extension-dapp": "^0.43.1",
     "@polkadot/networks": "^9.1.1",
     "@polkadot/react-identicon": "^2.1.1",
-    "@polkadot/util": "^9.1.1",
     "@polkadot/util-crypto": "^9.1.1",
+    "@polkadot/util": "^9.1.1",
     "@polkadot/wasm-crypto-wasm": "^6.0.1",
     "@polkadot/x-randomvalues": "^9.1.1",
     "@tanem/react-nprogress": "^5.0.0",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
     "bn.js": "^5.2.0",
+    "next-i18next": "^11.0.0",
+    "react-dom": "18.1.0",
+    "react": "18.1.0",
     "rxjs": "^7.2.0",
     "typescript": "4.6.3",
     "web3": "^1.7.3"
